Move all image persistence logic to the store

+++ memegen/stores/image.py
@@ -5,8 +5,9 @@ class ImageStore:
 
     LATEST = "latest.jpg"
 
-    def __init__(self, root):
+    def __init__(self, root, config):
         self.root = root
+        self.debug = config.get('DEBUG', False)
 
     @property
     def latest(self):
@@ -18,8 +19,12 @@ class ImageStore:
         return os.path.isfile(image.path) and not image.style
 
     def create(self, image):
+        if self.exists(image) and not self.debug:
+            return
+
         image.root = self.root
         image.generate()
+
         try:
             os.remove(self.latest)
         except IOError:
